semantic_search/requirements.txt
voyageai
pinecone
PyMuPDF
Pillow
flask
firebase-admin
gunicorn
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pytest
